refactor: offload mod.rs
